refactor(server): extract FakeLightningBackend into a shared test double.
